requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
python-dotenv==0.21.1
psycopg2==2.9.9
SQLAlchemy==2.0.36
